UdaConnect/modules/location_service_consumer/requirements.txt
flask==2.0.3
flask-restx==0.5.1
flask-cors==3.1.1
grpcio==1.43.0
grpcio-tools==1.43.0
kafka-python==2.0.2
psycopg2-binary==2.9.3
protobuf==3.19.0
